get_article/requirements.txt
google-cloud-storage==1.10.0
newspaper3k